added comments for main class and payable interface and fixed int conversion in compareTo method in person class.

+++ src/modules/Person.java
@@ -47,7 +47,7 @@ abstract public class Person implements Payable, Comparable<Person>{
 
     @Override
     public int compareTo(Person o) {  // This method lets you compare salaries of two people
-        return Integer.compare((int) this.getPaymentAmount(), (int) o.getPaymentAmount());
+        return Double.compare(this.getPaymentAmount(), o.getPaymentAmount());
     }
 
     @Override
